feat: Adding Device Simulator Panel
This is infrastructure for #63

So far, we have all the plumbing for getting messages to and from the C#
CLI from the frontend, the simulator panel is hidden by default and
a PoC "Add device" button adds a duel vibe which prints its commands to
the console.

More work is needed to make the devices customisable and implement
visualisations of the movements.

Ideally there'd also 2 default devices ready to 'connect': the Hush and
Launch from the original JS simulator.